+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Renamed the `Async` module to `Threading`.
 - Renamed the `Cache` and `CacheSerializer` interfaces to `ICache` and `ICacheSerializer`.
 - Added the `ICache.getOrCreate()` method.
+- Added the `isCompleted`, `isCompletedSuccessfully` and `isFaulted` properties to the `Task` class.
 
 ## Version [3.3.0](https://github.com/cedx/core.js/compare/v3.2.0...v3.3.0)
 - Removed the `DateRange` class.